Improved performance & integrate PolyExp Engine into StringConfusion
git-svn-id: https://Confuser.svn.codeplex.com/svn@48832 a83551a4-30f6-4d81-a974-c6ced450ddbf

+++ Confuser.Core/Poly/Expressions/ConstantExpression.cs
@@ -7,8 +7,8 @@ namespace Confuser.Core.Poly.Expressions
 {
     public class ConstantExpression : Expression
     {
-        int val;
-        public int Value { get { return val; } set { val = value; } }
+        double val;
+        public double Value { get { return val; } set { val = value; } }
 
         public override Expression GetVariableExpression()
         {
